initial refactor. allow optional dependencies to be loaded on demand
