+++ .pre-commit-config.yaml
@@ -1,13 +1,25 @@
+# lock file is not pretty json, but we shouldn't be messing with it
+exclude: 'Pipfile.lock'
 repos:
-  -   repo: https://github.com/python/black
-      rev: stable
-      hooks:
-      - id: black
-  -   repo: git://github.com/pre-commit/pre-commit-hooks
-      rev: v1.2.3
-      hooks:
-      - id: flake8
-  -   repo: https://github.com/pre-commit/mirrors-mypy
-      rev: v0.711
-      hooks:
+  - repo: https://github.com/python/black
+    rev: stable
+    hooks:
+    - id: black
+      language_version: python3.7
+  - repo: https://github.com/pre-commit/mirrors-mypy
+    rev: v0.782
+    hooks:
       - id: mypy
+  - repo: https://gitlab.com/pycqa/flake8
+    rev: 3.8.3
+    hooks:
+    - id: flake8
+  - repo: https://github.com/pre-commit/pre-commit-hooks
+    rev: v1.2.3
+    hooks:
+    - id: trailing-whitespace
+    - id: end-of-file-fixer
+    - id: check-yaml
+    - id: check-ast
+    - id: check-json
+    - id: pretty-format-json
