+++ LICENSE
@@ -1,4 +1,4 @@
-Copyright 2019 Blackfynn, Inc
+Copyright 2020 Blackfynn, Inc
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
