+++ README.rst
@@ -11,7 +11,8 @@
 Pathman
 =======
 
-Pathman is a utility for interacting with files using a uniform interface regardless of where
-the files live: locally, on S3, or in some other remote service. The interface attempts closely
-follow the `pathlib` interface when possible. Support for interacting with files in S3 is done 
-through heavy use of the s3fs_ library.
+Pathman is a utility for interacting with files using a uniform interface
+regardless of where the files live: locally, on S3, or in some other remote
+service. The interface attempts closely follow the `pathlib` interface when
+possible. Support for interacting with files in S3 is done through heavy use of
+the s3fs_ library.
