+++ pathman/_impl/__init__.py
@@ -1,0 +1,4 @@
+"""Backend-specific implementation of the pathman.Path interface"""
+from .blackfynn import BlackfynnPath
+from .s3 import S3Path
+from .local import LocalPath
